backend/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
psycopg2-binary
Werkzeug==3.1.3
